% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{sync_with_cran}
\alias{sync_with_cran}
\alias{sync_full_universe}
\alias{delete_package}
\alias{post_failure}
\alias{post_package}
\alias{put_package}
\alias{crantype}
\alias{db_packages}
\alias{db_files}
\title{cranlike tools}
\usage{
sync_with_cran(
  pkgs,
  types = c("src", "win", "mac", "oldwin", "oldmac"),
  dependencies = FALSE,
  user = "cran"
)

sync_full_universe(user)

delete_package(
  package,
  version = NULL,
  type = c("src", "win", "mac"),
  user = "cran"
)

post_failure(package, version, user = "cran")

post_package(
  path,
  package,
  version,
  type = c("src", "win", "mac"),
  user = "cran"
)

put_package(
  path,
  package,
  version,
  type = c("src", "win", "mac"),
  user = "cran"
)

crantype(type = c("src", "win", "mac", "oldwin", "oldmac"))

db_packages()

db_files()
}
\arguments{
\item{pkgs}{a vector of packages}

\item{types}{one or more of src, win, mac}

\item{dependencies}{also sync hard dependencies of pkgs}

\item{user}{name of the user or organization to publish}

\item{package}{name of a package}

\item{version}{optional string with version}

\item{type}{one of src, win, mac}

\item{path}{full path to file to upload}
}
\description{
Utilities for testing the CRANlike API.
}
